+++ uclibc/lib/contrib/uclibc/libm/float_wrappers.c
@@ -304,7 +304,7 @@ long_WRAPPER1(lround)
 float modff (float x, float *iptr)
 {
 	double y, result;
-	result = modf( x, &y );
+	result = modf( (double)x, &y );
 	*iptr = (float)y;
 	return (float) result;
 }
